+++ photon-model-adapters/vsphere/adapter/src/main/java/com/vmware/photon/controller/model/adapters/vsphere/VSphereEndpointAdapterService.java
@@ -16,20 +16,29 @@ package com.vmware.photon.controller.model.adapters.vsphere;
 import static com.vmware.photon.controller.model.adapterapi.EndpointConfigRequest.PRIVATE_KEYID_KEY;
 import static com.vmware.photon.controller.model.adapterapi.EndpointConfigRequest.PRIVATE_KEY_KEY;
 import static com.vmware.photon.controller.model.adapterapi.EndpointConfigRequest.REGION_KEY;
+import static com.vmware.xenon.common.Operation.STATUS_CODE_BAD_REQUEST;
 
 import java.net.URI;
-import java.util.concurrent.TimeUnit;
 import java.util.function.BiConsumer;
+import java.util.function.Consumer;
 
 import com.vmware.photon.controller.model.adapterapi.EndpointConfigRequest;
 import com.vmware.photon.controller.model.adapters.util.EndpointAdapterUtils;
 import com.vmware.photon.controller.model.adapters.util.EndpointAdapterUtils.Retriever;
+import com.vmware.photon.controller.model.adapters.vsphere.util.connection.BasicConnection;
+import com.vmware.photon.controller.model.adapters.vsphere.util.connection.ConnectionException;
+import com.vmware.photon.controller.model.adapters.vsphere.util.connection.GetMoRef;
+import com.vmware.photon.controller.model.adapters.vsphere.util.finders.Finder;
+import com.vmware.photon.controller.model.adapters.vsphere.util.finders.FinderException;
 import com.vmware.photon.controller.model.resources.ComputeDescriptionService.ComputeDescription;
 import com.vmware.photon.controller.model.resources.ComputeService.ComputeState;
+import com.vmware.vim25.InvalidPropertyFaultMsg;
+import com.vmware.vim25.RuntimeFaultFaultMsg;
 import com.vmware.xenon.common.Operation;
 import com.vmware.xenon.common.ServiceErrorResponse;
 import com.vmware.xenon.common.StatelessService;
 import com.vmware.xenon.common.UriUtils;
+import com.vmware.xenon.common.Utils;
 import com.vmware.xenon.services.common.AuthCredentialsService.AuthCredentialsServiceState;
 
 /**
@@ -58,19 +67,31 @@ public class VSphereEndpointAdapterService extends StatelessService {
 
         return (credentials, callback) -> {
             String host = body.endpointProperties.get(HOST_NAME_KEY);
-            VSphereIOThreadPool pool = VSphereIOThreadPool.createDefault(this.getHost(), 1);
 
             URI adapterManagementUri = getAdapterManagementUri(host);
-            pool.submit(this, adapterManagementUri, credentials, (connection, e) -> {
-                if (e != null) {
-                    String msg = String.format("Cannot establish connection to %s",
-                            adapterManagementUri);
-                    logInfo(msg);
-                    callback.accept(null, e);
-                } else {
-                    callback.accept(null, null);
+            String id = body.endpointProperties.get(REGION_KEY);
+
+            BasicConnection connection = createConnection(adapterManagementUri, credentials);
+            try {
+                // login and session creation
+                connection.connect();
+                if (id != null && !id.isEmpty()) {
+                    new Finder(connection, id);
                 }
-            }, (c) -> c.setRequestTimeout(5, TimeUnit.SECONDS));
+                callback.accept(null, null);
+            } catch (RuntimeFaultFaultMsg | InvalidPropertyFaultMsg | FinderException e) {
+                ServiceErrorResponse r = Utils.toServiceErrorResponse(e);
+                r.statusCode = STATUS_CODE_BAD_REQUEST;
+                r.message = String.format("Error looking for datacenter for id '%s'", id);
+                callback.accept(r, e);
+            } catch (ConnectionException e) {
+                String msg = String.format("Cannot establish connection to %s",
+                        adapterManagementUri);
+                logInfo(msg);
+                callback.accept(null, e);
+            } finally {
+                closeQuietly(connection);
+            }
         };
     }
 
@@ -84,17 +105,53 @@ public class VSphereEndpointAdapterService extends StatelessService {
 
     private BiConsumer<ComputeDescription, Retriever> computeDesc() {
         return (cd, r) -> {
-            cd.regionId = r.getRequired(REGION_KEY);
-            cd.zoneId = cd.regionId;
+            String id = r.getRequired(REGION_KEY);
+            Consumer<String> consumer = (regionId) -> {
+                cd.regionId = regionId;
+                cd.zoneId = cd.regionId;
 
-            cd.environmentName = ComputeDescription.ENVIRONMENT_NAME_ON_PREMISE;
-            cd.instanceAdapterReference = UriUtils.buildUri(getHost(),
-                    VSphereUriPaths.INSTANCE_SERVICE);
-            cd.enumerationAdapterReference = UriUtils.buildUri(getHost(),
-                    VSphereUriPaths.ENUMERATION_SERVICE);
-            cd.statsAdapterReference = UriUtils.buildUri(getHost(), VSphereUriPaths.STATS_SERVICE);
-            cd.powerAdapterReference = UriUtils.buildUri(getHost(), VSphereUriPaths.POWER_SERVICE);
+                cd.environmentName = ComputeDescription.ENVIRONMENT_NAME_ON_PREMISE;
+                cd.instanceAdapterReference = UriUtils.buildUri(getHost(),
+                        VSphereUriPaths.INSTANCE_SERVICE);
+                cd.enumerationAdapterReference = UriUtils.buildUri(getHost(),
+                        VSphereUriPaths.ENUMERATION_SERVICE);
+                cd.statsAdapterReference = UriUtils.buildUri(getHost(),
+                        VSphereUriPaths.STATS_SERVICE);
+                cd.powerAdapterReference = UriUtils.buildUri(getHost(),
+                        VSphereUriPaths.POWER_SERVICE);
+            };
+
+            if (id == null || id.isEmpty()) {
+                loadDefault(r, consumer);
+            } else {
+                consumer.accept(id);
+            }
+
         };
+    }
+
+    private void loadDefault(Retriever r, Consumer<String> consumer) {
+
+        String host = r.getRequired(HOST_NAME_KEY);
+        URI adapterManagementUri = getAdapterManagementUri(host);
+        AuthCredentialsServiceState ac = new AuthCredentialsServiceState();
+        credentials().accept(ac, r);
+
+        BasicConnection connection = createConnection(adapterManagementUri, ac);
+        try {
+            // login and session creation
+            connection.connect();
+            Finder finder = new Finder(connection, null);
+            String datacenterName = new GetMoRef(connection)
+                    .entityProp(finder.getDatacenter().object, "name");
+            consumer.accept(datacenterName);
+        } catch (RuntimeFaultFaultMsg | InvalidPropertyFaultMsg | FinderException ex) {
+            throw new IllegalArgumentException("Error looking for default datacenter", ex);
+        } catch (ConnectionException e) {
+            throw new IllegalArgumentException(e);
+        } finally {
+            closeQuietly(connection);
+        }
     }
 
     private BiConsumer<ComputeState, Retriever> compute() {
@@ -108,5 +165,28 @@ public class VSphereEndpointAdapterService extends StatelessService {
         vcUrl.append(host);
         vcUrl.append("/sdk");
         return UriUtils.buildUri(vcUrl.toString());
+    }
+
+    private BasicConnection createConnection(URI adapterReference,
+            AuthCredentialsServiceState auth) {
+        BasicConnection connection = new BasicConnection();
+
+        // TODO control sslErrors policy externally
+        connection.setIgnoreSslErrors(true);
+
+        connection.setUsername(auth.privateKeyId);
+        connection.setPassword(auth.privateKey);
+
+        connection.setURI(adapterReference);
+
+        return connection;
+    }
+
+    private void closeQuietly(BasicConnection connection) {
+        try {
+            connection.close();
+        } catch (Exception e) {
+            logWarning("Error closing connection to " + connection.getURI(), e);
+        }
     }
 }