+++ photon-model/src/main/java/com/vmware/photon/controller/model/resources/ComputeDescriptionService.java
@@ -91,12 +91,6 @@ public class ComputeDescriptionService extends StatefulService {
         public String environmentName;
 
         /**
-         * Identifier of the data center associated with this compute host.
-         */
-        @UsageOption(option = PropertyUsageOption.AUTO_MERGE_IF_NOT_NULL)
-        public String datacenterId;
-
-        /**
          * Identifier of the data store associated with this compute host.
          */
         @UsageOption(option = PropertyUsageOption.AUTO_MERGE_IF_NOT_NULL)
@@ -342,7 +336,6 @@ public class ComputeDescriptionService extends StatefulService {
         template.enumerationAdapterReference = UriUtils.buildUri(
                 this.getHost(), "/enumerationAdapterReference");
 
-        template.datacenterId = null;
         template.networkId = null;
         template.dataStoreId = null;